+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -40,8 +40,10 @@ import EventsScope from '@webex/plugin-meetings/src/common/events/events-scope';
 import Meetings, {CONSTANTS} from '@webex/plugin-meetings';
 import Meeting from '@webex/plugin-meetings/src/meeting';
 import Members from '@webex/plugin-meetings/src/members';
+import * as MembersImport from '@webex/plugin-meetings/src/members';
 import Roap from '@webex/plugin-meetings/src/roap';
 import MeetingRequest from '@webex/plugin-meetings/src/meeting/request';
+import * as MeetingRequestImport from '@webex/plugin-meetings/src/meeting/request';
 import LocusInfo from '@webex/plugin-meetings/src/locus-info';
 import MediaProperties from '@webex/plugin-meetings/src/media/properties';
 import MeetingUtil from '@webex/plugin-meetings/src/meeting/util';
@@ -169,6 +171,8 @@ describe('plugin-meetings', () => {
   let test3;
   let test4;
   let testDestination;
+  let membersSpy;
+  let meetingRequestSpy;
 
   beforeEach(() => {
     webex = new MockWebex({
@@ -203,6 +207,8 @@ describe('plugin-meetings', () => {
     webex.internal.metrics.submitClientMetrics = sinon.stub().returns(Promise.resolve());
     webex.meetings.uploadLogs = sinon.stub().returns(Promise.resolve());
     webex.internal.llm.on = sinon.stub();
+    membersSpy = sinon.spy(MembersImport, 'default');
+    meetingRequestSpy = sinon.spy(MeetingRequestImport, 'default');
 
     TriggerProxy.trigger = sinon.stub().returns(true);
     Metrics.postEvent = sinon.stub();
@@ -256,6 +262,16 @@ describe('plugin-meetings', () => {
           assert.equal(meeting.deviceUrl, uuid3);
           assert.deepEqual(meeting.meetingInfo, {});
           assert.instanceOf(meeting.members, Members);
+          assert.calledOnceWithExactly(
+            membersSpy,
+            {
+              locusUrl: meeting.locusUrl,
+              receiveSlotManager: meeting.receiveSlotManager,
+              mediaRequestManagers: meeting.mediaRequestManagers,
+              meeting,
+            },
+            {parent: meeting.webex}
+          );
           assert.instanceOf(meeting.roap, Roap);
           assert.instanceOf(meeting.reconnectionManager, ReconnectionManager);
           assert.isNull(meeting.audio);
@@ -270,6 +286,13 @@ describe('plugin-meetings', () => {
           assert.isNull(meeting.hostId);
           assert.isNull(meeting.policy);
           assert.instanceOf(meeting.meetingRequest, MeetingRequest);
+          assert.calledOnceWithExactly(
+            meetingRequestSpy,
+            {
+              meeting,
+            },
+            {parent: meeting.webex}
+          );
           assert.instanceOf(meeting.locusInfo, LocusInfo);
           assert.equal(meeting.fetchMeetingInfoTimeoutId, undefined);
           assert.instanceOf(meeting.mediaProperties, MediaProperties);
@@ -2963,7 +2986,6 @@ describe('plugin-meetings', () => {
           meeting.meetingRequest.dialOut = sinon
             .stub()
             .returns(Promise.resolve({body: {locus: 'testData'}}));
-          meeting.locusInfo.onFullLocus = sinon.stub().returns(Promise.resolve());
         });
 
         it('with no parameters triggers dial-in, delegating request to meetingRequest correctly', async () => {
@@ -2976,11 +2998,9 @@ describe('plugin-meetings', () => {
             locusUrl: meeting.locusUrl,
             clientUrl: meeting.deviceUrl,
           });
-          assert.calledWith(meeting.locusInfo.onFullLocus, 'testData');
           assert.notCalled(meeting.meetingRequest.dialOut);
 
           meeting.meetingRequest.dialIn.resetHistory();
-          meeting.locusInfo.onFullLocus.resetHistory();
 
           // try again. the dial in urls should match
           await meeting.usePhoneAudio();
@@ -2991,7 +3011,6 @@ describe('plugin-meetings', () => {
             locusUrl: meeting.locusUrl,
             clientUrl: meeting.deviceUrl,
           });
-          assert.calledWith(meeting.locusInfo.onFullLocus, 'testData');
           assert.notCalled(meeting.meetingRequest.dialOut);
         });
 
@@ -3008,11 +3027,9 @@ describe('plugin-meetings', () => {
             clientUrl: meeting.deviceUrl,
             phoneNumber,
           });
-          assert.calledWith(meeting.locusInfo.onFullLocus, 'testData');
           assert.notCalled(meeting.meetingRequest.dialIn);
 
           meeting.meetingRequest.dialOut.resetHistory();
-          meeting.locusInfo.onFullLocus.resetHistory();
 
           // try again. the dial out urls should match
           await meeting.usePhoneAudio(phoneNumber);
@@ -3024,7 +3041,6 @@ describe('plugin-meetings', () => {
             clientUrl: meeting.deviceUrl,
             phoneNumber,
           });
-          assert.calledWith(meeting.locusInfo.onFullLocus, 'testData');
           assert.notCalled(meeting.meetingRequest.dialIn);
         });
 
@@ -4110,6 +4126,43 @@ describe('plugin-meetings', () => {
       });
     });
 
+    describe('#enableMusicMode', () => {
+      beforeEach(() => {
+        meeting.isMultistream = true;
+          meeting.mediaProperties.webrtcMediaConnection = {
+            setCodecParameters: sinon.stub().resolves({}),
+            deleteCodecParameters: sinon.stub().resolves({}),
+          };
+      });
+      [
+        {shouldEnableMusicMode: true},
+        {shouldEnableMusicMode: false},
+      ].forEach(({shouldEnableMusicMode}) => {
+        it(`fails if there is no media connection for shouldEnableMusicMode: ${shouldEnableMusicMode}`, async () => {
+          meeting.mediaProperties.webrtcMediaConnection = undefined;
+          await assert.isRejected(meeting.enableMusicMode(shouldEnableMusicMode));
+        });
+      });
+
+      it('should set the codec parameters when shouldEnableMusicMode is true', async () => {
+        await meeting.enableMusicMode(true);
+        assert.calledOnceWithExactly(meeting.mediaProperties.webrtcMediaConnection.setCodecParameters, MediaType.AudioMain, {
+          maxaveragebitrate: '64000',
+          maxplaybackrate: '48000',
+        });
+        assert.notCalled(meeting.mediaProperties.webrtcMediaConnection.deleteCodecParameters);
+      });
+
+      it('should set the codec parameters when shouldEnableMusicMode is false', async () => {
+        await meeting.enableMusicMode(false);
+        assert.calledOnceWithExactly(meeting.mediaProperties.webrtcMediaConnection.deleteCodecParameters, MediaType.AudioMain, [
+          'maxaveragebitrate',
+          'maxplaybackrate',
+        ]);
+        assert.notCalled(meeting.mediaProperties.webrtcMediaConnection.setCodecParameters);
+      });
+    });
+
     describe('Public Event Triggers', () => {
       let sandbox;
       const {ENDED} = CONSTANTS;
@@ -4947,6 +5000,24 @@ describe('plugin-meetings', () => {
           );
         });
 
+        it('listens to MEETING_CONTROLS_VIDEO_UPDATED', async () => {
+          const state = {example: 'value'}
+
+          await meeting.locusInfo.emitScoped(
+            {function: 'test', file: 'test'},
+            LOCUSINFO.EVENTS.CONTROLS_VIDEO_CHANGED,
+            {state}
+          );
+
+          assert.calledWith(
+            TriggerProxy.trigger,
+            meeting,
+            {file: 'meeting/index', function: 'setupLocusControlsListener'},
+            EVENT_TRIGGERS.MEETING_CONTROLS_VIDEO_UPDATED,
+            {state},
+          );
+        });
+
         it('listens to the timing that user joined into breakout', async () => {
           const mainLocusUrl = 'mainLocusUrl123';
 
@@ -5376,28 +5447,7 @@ describe('plugin-meetings', () => {
           checkParseMeetingInfo(expectedInfoToParse);
         });
       });
-      describe('#parseLocus', () => {
-        describe('when CALL and participants', () => {
-          beforeEach(() => {
-            meeting.setLocus = sinon.stub().returns(true);
-            MeetingUtil.getLocusPartner = sinon.stub().returns({person: {sipUrl: uuid3}});
-          });
-          it('should parse the locus object and set meeting properties and return null', () => {
-            meeting.type = 'CALL';
-            meeting.parseLocus({url: url1, participants: [{id: uuid1}], self: {id: uuid2}});
-            assert.calledOnce(meeting.setLocus);
-            assert.calledWith(meeting.setLocus, {
-              url: url1,
-              participants: [{id: uuid1}],
-              self: {id: uuid2},
-            });
-            assert.calledOnce(MeetingUtil.getLocusPartner);
-            assert.calledWith(MeetingUtil.getLocusPartner, [{id: uuid1}], {id: uuid2});
-            assert.deepEqual(meeting.partner, {person: {sipUrl: uuid3}});
-            assert.equal(meeting.sipUri, uuid3);
-          });
-        });
-      });
+
       describe('#setCorrelationId', () => {
         it('should set the correlationId and return undefined', () => {
           assert.ok(meeting.correlationId);
@@ -5748,8 +5798,9 @@ describe('plugin-meetings', () => {
         beforeEach(() => {
           meeting.locusInfo.initialSetup = sinon.stub().returns(true);
         });
+
         it('should read the locus object, set on the meeting and return null', () => {
-          meeting.parseLocus({
+          meeting.setLocus({
             mediaConnections: [test1],
             locusUrl: url1,
             locusId: uuid1,
@@ -5774,6 +5825,7 @@ describe('plugin-meetings', () => {
           assert.equal(meeting.hostId, uuid4);
         });
       });
+
       describe('preferred video device', () => {
         describe('#getVideoDeviceId', () => {
           it('returns the preferred video device', () => {
@@ -5863,7 +5915,7 @@ describe('plugin-meetings', () => {
               'https://board-a.wbx2.com/board/api/v1/channels/977a7330-54f4-11eb-b1ef-91f5eefc7bf3',
           };
 
-          const generateContent = (beneficiaryId = null, disposition = null) => ({
+          const generateContent = (beneficiaryId = null, disposition = null,annotation = undefined) => ({
             beneficiaryId,
             disposition,
           });
@@ -5880,7 +5932,10 @@ describe('plugin-meetings', () => {
             beneficiaryId,
             resourceUrl,
             isAccepting,
-            otherBeneficiaryId
+            otherBeneficiaryId,
+            annotation,
+            url,
+            shareInstanceId
           ) => {
             const newPayload = cloneDeep(payload);
 
@@ -5906,7 +5961,7 @@ describe('plugin-meetings', () => {
             if (isGranting) {
               if (isContent) {
                 activeSharingId.content = beneficiaryId;
-                newPayload.current.content = generateContent(beneficiaryId, FLOOR_ACTION.GRANTED);
+                newPayload.current.content = generateContent(beneficiaryId, FLOOR_ACTION.GRANTED,annotation);
 
                 if (isEqual(newPayload.current, newPayload.previous)) {
                   eventTrigger.member = null;
@@ -5960,7 +6015,7 @@ describe('plugin-meetings', () => {
                     eventTrigger.share.push({
                       eventName: EVENT_TRIGGERS.MEETING_STARTED_SHARING_REMOTE,
                       functionName: 'remoteShare',
-                      eventPayload: {memberId: beneficiaryId},
+                      eventPayload: {memberId: beneficiaryId, url, shareInstanceId},
                     });
                   }
                 }
@@ -6530,6 +6585,28 @@ describe('plugin-meetings', () => {
               const data3 = generateData(data2.payload, false, false, USER_IDS.REMOTE_A);
 
               payloadTestHelper([data1, data2, data3]);
+            });
+          });
+
+          describe('annotation policy', () => {
+
+            it('Scenario #1: blank annotation', () => {
+              const data1 = generateData(blankPayload, true, true, USER_IDS.ME);
+              const data2 = generateData(data1.payload, false, true, USER_IDS.ME);
+              const data3 = generateData(data2.payload, true, true, USER_IDS.ME);
+              const data4 = generateData(data3.payload, false, true, USER_IDS.ME);
+
+              payloadTestHelper([data1, data2, data3, data4]);
+            });
+
+            it('Scenario #2: annotation', () => {
+              const annotationInfo = {version: '1', policy: 'Approval'};
+              const data1 = generateData(blankPayload, true, true, USER_IDS.ME, annotationInfo);
+              const data2 = generateData(data1.payload, false, true, USER_IDS.ME);
+              const data3 = generateData(data2.payload, true, true, USER_IDS.ME);
+              const data4 = generateData(data3.payload, false, true, USER_IDS.ME);
+
+              payloadTestHelper([data1, data2, data3, data4]);
             });
           });